Pipelines/V0/Orchestrator/pipelines-orc-orchestratorpipeline-v0/bin/requirements.txt
pyyaml
requests
awesome-slugify